[Fix] 배열 길이 인자 타입 size_t로 수정, 헤더 추가

+++ Module07/ex01/iter.hpp
@@ -1,9 +1,11 @@
 #ifndef ITER_HPP
 # define ITER_HPP
 
+# include <iostream>
+
 template <typename T>
-void iter(T *array, unsigned int length, void (*f)(T const &)) {
-	for (unsigned int i = 0; i < length; i++) {
+void iter(T *array, size_t length, void (*f)(T const &)) {
+	for (size_t i = 0; i < length; i++) {
 		f(array[i]);
 	}
 };
